Wrote scraper for residential restaurants and The Study
I learned the hard way that the dining website's HTML is very
inconsistent between different dining halls/restaurants. A separate
function will have to be written for each of them. Thankfully, there is
some commonality in how the "menu-item" containers are structured, so I
can use one function to extract information from each item. The scrapers
for other venues will not be hard to implement.